pyproject.toml
[project]
name = "cad-agent"
version = "0.1.0"
description = "Agentic pipeline: prompt/image/video to parametric CAD and STEP"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "python-multipart>=0.0.9",
  "orjson>=3.10.0",
  "jsonschema>=4.22.0",
  "rich>=13.7.0",
  "pyyaml>=6.0.1",
  "requests>=2.32.0",
  "huggingface_hub[cli]>=0.24.0",
  "trimesh>=4.4.0",
  "numpy>=1.26.0",
  "shapely>=2.0.0",
  "cadquery>=2.4.0",
  "tenacity>=8.3.0"
]

[tool.ruff]
line-length = 100

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
